fix: address billing flow review feedback
